requirements.txt
# LLM and AI
openai>=1.0.0
langchain-openai>=0.1.0
langgraph>=0.0.20
google-genai>=1.0.0

# Vector Store
pinecone>=3.0.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# API Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# HTTP Client
httpx>=0.25.0

# Document Loader
pypdf>=5.0.0

# ENV
python-dotenv>=1.0.0

# Redis
redis>=5.0.0